Sonar me please

+++ DAO/src/main/java/com/netcracker/hb/Dao/CRUD/hotel/FloorCRUD.java
@@ -1,12 +1,12 @@
 package com.netcracker.hb.Dao.CRUD.hotel;
 
 import com.netcracker.hb.Dao.CRUD.CRUD;
+import com.netcracker.hb.Dao.CRUD.DatabaseProperties;
 import com.netcracker.hb.entities.hotel.Floor;
 import com.netcracker.hb.entities.hotel.Hotel;
 import com.netcracker.hb.entities.hotel.Room;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
@@ -20,11 +20,20 @@ import lombok.extern.log4j.Log4j;
 @Log4j
 public class FloorCRUD implements CRUD<Floor> {
 
-  private static final CRUD<Floor> floorCRUD = new FloorCRUD();
-  private FloorCRUD(){}
-  public static CRUD<Floor> getFloorCRUD(){
+  private static CRUD<Floor> floorCRUD;
+  private FloorCRUD() {
+  }
+  public static synchronized CRUD<Floor> getFloorCRUD() {
+    if(floorCRUD ==null){
+      floorCRUD = new FloorCRUD();
+    }
     return floorCRUD;
   }
+  private static final String START = "<Start searching floors....";
+  private static final String END = "Floors was found>";
+  private static final String ERROR = "Floors was not found>";
+
+
 
   private static final CRUD<Hotel> hotelCRUD = HotelCRUD.getHotelCRUD();
   private static final CRUD<Room> roomsCRUD = RoomsCRUD.getRoomsCRUD();
@@ -32,139 +41,120 @@ public class FloorCRUD implements CRUD<Floor> {
 
   @Override
   public List<Floor> searchObjects() {
-    log.info("<Start searching floors....");
-    File floorFolderDirectory = new File("entSAVE/floor_entities");
+    log.info(START);
+    File floorFolderDirectory = new File(DatabaseProperties.getFloorCrudEntitiesPath());
     String[] floorList = floorFolderDirectory.list();
-    List floors = new ArrayList();
-    try {
-      for (String floorFolderName : floorList) {
-
-        FileInputStream fileFloorIn = new FileInputStream("entSAVE/floor_entities/"
-            + floorFolderName);
-        ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn);
+    List<Floor> floors = new ArrayList<>();
+    assert floorList != null;
+    for (String floorFolderName : floorList) {
+      try (
+          FileInputStream fileFloorIn = new FileInputStream(
+              DatabaseProperties.getFloorCrudEntitiesPath() + floorFolderName);
+          ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn)) {
         Floor object = (Floor) objectFloorIn.readObject();
-        //если номер этажа совпадает то передаем
         floors.add(object);
-        objectFloorIn.close();
+      } catch (Exception exception) {
+        exception.printStackTrace();
       }
-    } catch (FileNotFoundException exception) {
-      exception.printStackTrace();
-    } catch (IOException e) {
-      e.printStackTrace();
-    } catch (ClassNotFoundException e) {
-      e.printStackTrace();
-    } finally {
-      if(floors.isEmpty()){
-        log.error("FLOORS NOT FOUND>");
-      }
-      return floors;
     }
+    if (floors.isEmpty()) {
+      log.error(ERROR);
+    }
+    return floors;
   }
 
   @Override
   public Floor searchObjectNum(int floorNum) {
 
-    log.info("<Start searching floor....");
-    File floorFolderDirectory = new File("entSAVE/floor_entities");
+    log.info(START);
+    File floorFolderDirectory = new File(DatabaseProperties.getFloorCrudEntitiesPath());
     String[] floorList = floorFolderDirectory.list();
     Floor floor = null;
-    try {
-      for (String floorFolderName : floorList) {
-
-        FileInputStream fileFloorIn = new FileInputStream("entSAVE/floor_entities/"
-            + floorFolderName);
-        ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn);
+    assert floorList != null;
+    for (String floorFolderName : floorList) {
+      try (
+          FileInputStream fileFloorIn = new FileInputStream(
+              DatabaseProperties.getFloorCrudEntitiesPath() + floorFolderName);
+          ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn)) {
         Floor object = (Floor) objectFloorIn.readObject();
         //если номер этажа совпадает то передаем
         if (object.getFloorNum() == floorNum) {
-          log.info("Floor was found>");
+          log.info(END);
           floor = object;
         }
-        objectFloorIn.close();
+      } catch (IOException | ClassNotFoundException exception) {
+        exception.printStackTrace();
       }
-    } catch (FileNotFoundException exception) {
-      exception.printStackTrace();
-    } catch (IOException e) {
-      e.printStackTrace();
-    } catch (ClassNotFoundException e) {
-      e.printStackTrace();
-    } finally {
-      if(floor ==null){
-        log.error("FLOOR NOT FOUND>");
-      }
-      return floor;
     }
+    if (floor == null) {
+      log.error(ERROR);
+    }
+    return floor;
 
   }
 
   @Override
   public Floor searchUUIDObject(UUID uuid) {
-    log.info("<Start searching floor....");
-    File floorFolderDirectory = new File("entSAVE/floor_entities");
+    log.info(START);
+    if(uuid == null){
+      return null;
+    }
+    File floorFolderDirectory = new File(DatabaseProperties.getFloorCrudEntitiesPath());
     String[] floorList = floorFolderDirectory.list();
     Floor floor = null;
-    try {
-      for (String floorFolderName : floorList) {
 
-        FileInputStream fileFloorIn = new FileInputStream("entSAVE/floor_entities/"
-            + floorFolderName);
-        ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn);
+    assert floorList != null;
+    for (String floorFolderName : floorList) {
+      try (
+          FileInputStream fileFloorIn = new FileInputStream(
+              DatabaseProperties.getFloorCrudEntitiesPath() + floorFolderName);
+          ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn)) {
         Floor object = (Floor) objectFloorIn.readObject();
-        //если номер этажа совпадает то передаем
         if (object.getUuid().equals(uuid)) {
-          log.info("Floor was found>");
+          log.info(END);
           floor = object;
         }
-        objectFloorIn.close();
+      } catch (IOException |
+          ClassNotFoundException exception) {
+        exception.printStackTrace();
       }
-    } catch (FileNotFoundException exception) {
-      exception.printStackTrace();
-    } catch (IOException e) {
-      e.printStackTrace();
-    } catch (ClassNotFoundException e) {
-      e.printStackTrace();
-    } finally {
-      if(floor ==null){
-        log.error("FLOOR NOT FOUND>");
-      }
-      return floor;
     }
+    if (floor == null) {
+      log.error(ERROR);
+    }
+    return floor;
   }
 
   @Override
   public String searchFileName(Floor floor) {
 
-    log.info("<Start searching file name of floor...");
+    log.info(START);
 
-    File floorFolderDirectory = new File("entSAVE/floor_entities");
+    File floorFolderDirectory = new File(DatabaseProperties.getFloorCrudEntitiesPath());
     String[] floorList = floorFolderDirectory.list();
     String fileName = null;
-    try {
-      for (String floorFolderName : floorList) {
 
-        FileInputStream fileFloorIn = new FileInputStream("entSAVE/floor_entities/"
-            + floorFolderName);
-        ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn);
+    assert floorList != null;
+    for (String floorFolderName : floorList) {
+      try (
+          FileInputStream fileFloorIn = new FileInputStream(
+              DatabaseProperties.getFloorCrudEntitiesPath() + floorFolderName);
+          ObjectInputStream objectFloorIn = new ObjectInputStream(fileFloorIn)) {
         Floor object = (Floor) objectFloorIn.readObject();
 
         if (object.equals(floor)) {
-          log.info("File name of floor was found>");
+          log.info(END);
           fileName = floorFolderName;
         }
-        objectFloorIn.close();
+
+      } catch (IOException | ClassNotFoundException exception) {
+        exception.printStackTrace();
       }
-    } catch (FileNotFoundException exception) {
-      exception.printStackTrace();
-    } catch (IOException e) {
-      e.printStackTrace();
-    } catch (ClassNotFoundException e) {
-      e.printStackTrace();
-    } finally {
-      if(fileName == null){
-        log.error("FILE NAME OF FLOOR NOT FOUND>");
-      }
-      return fileName;
     }
+    if (fileName == null) {
+      log.error(ERROR);
+    }
+    return fileName;
 
   }
 
@@ -173,30 +163,28 @@ public class FloorCRUD implements CRUD<Floor> {
     log.info("<Start delete floor...");
 
     Floor floor = searchObjectNum(object.getFloorNum());
+    if (floor != null) {
 
-    //отвязываем от отеля
-    Hotel hotel = hotelCRUD.searchUUIDObject(floor.getUuid());
-    hotel.deleteFloor(floor.getUuid());
-    hotelCRUD.saveObject(hotel);
+      //отвязываем от отеля
+      Hotel hotel = hotelCRUD.searchUUIDObject(floor.getUuid());
+      hotel.deleteFloor(floor.getUuid());
+      hotelCRUD.saveObject(hotel);
 
-    //удаляем все внутренние комнаты
-    if (floor.getRoomsID() != null) {
-      for (UUID uuid : floor.getRoomsID()) {
-        roomsCRUD.deleteObject(roomsCRUD.searchUUIDObject(uuid));
+      //удаляем все внутренние комнаты
+      if (floor.getRoomsID() != null) {
+        for (UUID uuid : floor.getRoomsID()) {
+          roomsCRUD.deleteObject(roomsCRUD.searchUUIDObject(uuid));
+        }
       }
-    }
-    //Сохраняем
-    floor = searchObjectNum(object.getFloorNum());
-
-    //todo возможно прийдется удалять еще привязанных гостей
-    //удаляем файл
-    File deleteFile = new File("entSAVE/floor_entities/" + searchFileName(floor));
-
-    if (deleteFile != null) {
-      log.info("Floor was successfully deleted>" );
-      deleteFile.delete();
-    } else{
-      log.warn("NOTHING WAS DELETED, FILE FLOOR NOT FOUND>");
+      //Сохраняем
+      floor = searchObjectNum(object.getFloorNum());
+      //удаляем файл
+      File deleteFile = new File(DatabaseProperties.getFloorCrudEntitiesPath() + searchFileName(floor));
+      if (deleteFile.delete()) {
+        log.info("Floor was successfully deleted>");
+      }
+    } else {
+      log.warn(ERROR);
     }
   }
 
@@ -204,13 +192,12 @@ public class FloorCRUD implements CRUD<Floor> {
   public void saveObject(Floor floor) {
     log.info("<Start saving floor...");
 
-    try {
-      FileOutputStream fileFloorOut = new FileOutputStream("entSAVE/floor_entities/"
-           + floor.getUuid() + "-floor.txt");
+    try (
+        FileOutputStream fileFloorOut = new FileOutputStream(
+            DatabaseProperties.getFloorCrudEntitiesPath() + floor.getUuid() + "-floor.txt");
 
-      ObjectOutputStream objectFloorOut = new ObjectOutputStream(fileFloorOut);
+        ObjectOutputStream objectFloorOut = new ObjectOutputStream(fileFloorOut)) {
       objectFloorOut.writeObject(floor);
-      objectFloorOut.close();
       log.info("Success saving floor>");
     } catch (Exception ex) {
       ex.printStackTrace();
